did stuff to dvector and created dlist and dnode - oh and readmes
vector now uses smart pointers, specfically unique_ptr to use as a container instead of a raw pointer to make an array.
created the dnode class that will be used to for the dlink. the dnode class is pretty much set up for a doubly linked list.
created the dlist class. implemented the constructor, destructor, empty, size, front, back and push_front functions. at the moment, destructor is working how i think it is working. push_front is doing good things in world record speed. i am hoping a pizza will be in my belly in the near future.