queue array seems finished.
queue array seems finished. i added in main how i tested it. i pretty much just grew it and shrunk it a lot and checked if the queue lines up properly with how it communicates with the. the tricky part is keeping track of front and back when it is dynamically changing.